requirements.txt
alembic==1.14.0
fastapi==0.115.6
loguru==0.7.2
pydantic==2.10.3
python-dotenv==1.0.1
SQLAlchemy==2.0.36
uvicorn-loguru-integration==0.3.0
psycopg2-binary==2.9.10
authx==1.4.0
asyncpg==0.30.0
bcrypt==4.2.1
sqladmin==0.20.1
